requirements.txt
pandas==2.3.1
numpy==2.3.2
matplotlib==3.10.1
seaborn==0.13.2
plotly==6.3.0
wordcloud==1.9.4
streamlit==1.41.1
altair==5.5.0
boto3==1.35.25
s3fs==2025.7.0
pyathena==3.9.0  
python-dotenv==1.0.1
jupyter==1.1.1
ipykernel==6.29.5
notebook==7.2.2
